locksum_finance_full/backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
psycopg2-binary
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
plaid-python>=15.0.0
stripe>=7,<12
email-validator>=2.0.0
